Fix Adding validations

+++ api/src/services/campusService.js
@@ -1,3 +1,4 @@
+const ValidationError = require("../errors/ValidationError");
 const Campus = require("../models/Campus");
 const mongoose = require("mongoose");
 
@@ -11,44 +12,52 @@ const getCampuses = async (where = {}, skip, limit) => {
   const campuses = await Campus.find(where).skip(skip).limit(limit);
 
   return campuses;
-}
+};
 
 const getCampusById = async (id) => {
   const campus = await Campus.findById(id).exec();
 
   return campus;
-}
+};
 
 const getCountCampuses = async (where = {}) => {
   const numberCampuses = await Campus.count(where);
 
   return numberCampuses;
-}
+};
 
 const updateCampusById = async (id, campusData) => {
-  const updateCampus = await Campus.findByIdAndUpdate(id, campusData, { new: true });
+  const updateCampus = await Campus.findByIdAndUpdate(id, campusData, {
+    new: true,
+  });
 
   return updateCampus;
-}
+};
 
 const deleteCampusById = async (id) => {
+  if (!mongoose.isValidObjectId(id))
+    throw new ValidationError("El id debe ser un objectId");
+
   const campusDeleted = await Campus.findByIdAndDelete(id);
 
   return campusDeleted;
-}
+};
 
 const deleteCampus = async (where = {}) => {
+  if (Object.keys(where).lenght <= 0)
+    throw new ValidationError("Proporcione un criterio de filtrado");
+
   const numberCampusesDeleted = await Campus.deleteOne(where);
 
   return numberCampusesDeleted;
-}
+};
 
-module.exports = { 
-  createCampus, 
-  getCampuses, 
-  getCampusById, 
-  getCountCampuses, 
-  updateCampusById, 
-  deleteCampusById, 
-  deleteCampus 
+module.exports = {
+  createCampus,
+  getCampuses,
+  getCampusById,
+  getCountCampuses,
+  updateCampusById,
+  deleteCampusById,
+  deleteCampus,
 };
